move demoparse out of ghost subdir
