+++ trunk/ghost/ghostparse.c
@@ -19,7 +19,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 
 #include "../quakedef.h"
-#include "demoparse.h"
+#include "../demoparse.h"
 #include "ghost_private.h"
